Move font-family rule to [data-amplify-theme] (#1065)
* Move font-family rule to [data-amplify-theme]

This fixes an issue where defining a fonts token in custom theme didn't
override the default Inter font. This was occurring because the
`font-family: var(--amplify-fonts-default-static)` is only defined at the root. The AmplifyProvider converts custom theme tokens to CSS variables at the [data-amplify-theme] node, which means the custom theme font wasn't being picked up. Moving it to the [data-amplify-theme] ensures
that the custom font token is picked up correctly.