Use explicit aws-amplify imports for ui and react:
Removes the dependency on the umbrella `aws-amplify` package and
instead maps to @aws-amplify/core and @aws-amplify/auth packages.
The benefit of this change is to prevent importing other unused
packages which contain NodeJS globals. This makes it simpler
to use our react package with Vite/ESBuild, requiring customers only to
polyfill the `global` variable due to `@aws-amplify/auth` but not `process`
or `exports`.