chore (ci): update RN build tests to include standalone 0.78 release & add Expo/RN 0.77 test
RN 0.79 was recently released and the latest tests are no longer testing RN 0.78 actively.
However, Expo has not released or provided stable support for their SDK to support RN 0.78.
Furthermore, it was discovered that there was an extra check missed in order to test
RN 0.77 with Expo SDK 52, which meant we were testing RN 0.76 instead of RN 0.77.